+++ app/javascript/activeadmin_addons/inputs/slim-select-nested.js
@@ -12,7 +12,7 @@ function ajaxSearch(search, currentData, args) {
   });
 
   if (!!args.parent) {
-    args.query.q[`${args.parent}_eq`] = args.parentId;
+    args.query.q[`${args.parent}_id_eq`] = args.parentId;
   }
 
   args.query.q = { ...args.query.q, ...args.filters };
@@ -25,10 +25,12 @@ function collectionSearch(search, args, collection) {
     return Promise.reject('Search term too short');
   }
 
-  const data = JSON.parse(collection).map(
+  const data = JSON.parse(collection).filter(
+    (item) => (!args.parent || item[args.parent] === Number(args.parentId)) &&
+      String(item.text).toLowerCase().includes(search.toLowerCase()),
+  ).map(
     (item) => ({ value: String(item.id), text: item.text }),
-  ).filter(
-    (item) => String(item.text).toLowerCase().includes(search.toLowerCase()));
+  );
 
   return Promise.resolve(data);
 }
@@ -48,7 +50,7 @@ function settings(el) {
         const { model, association } = el.dataset;
         const child = el.closest('.nested_level')
           .parentNode
-          .querySelector(`.nested_level [data-model=${model}][data-parent=${association}_id]`);
+          .querySelector(`.nested_level [data-model=${model}][data-parent=${association}]`);
 
         if (child) {
           child.slim.setSelected();