+++ spec/features/inputs/tags_input_spec.rb
@@ -34,7 +34,7 @@ describe "Tags Input", type: :feature do
     end
   end
 
-  context "working with active record relations" do
+  context "when working with active record relations" do
     before do
       register_form(Invoice) do |f|
         f.input :item_ids, as: :tags, collection: Item.all
@@ -51,26 +51,26 @@ describe "Tags Input", type: :feature do
     context "with added item" do
       before { pick_slimselect_entered_option(@item1.name) }
 
-      it "adds/removes hidden item", js: true do
-        item_id = "#invoice_item_ids_#{@item1.id}"
-        input = find(item_id, visible: false)
-        expect(input.value).to eq(@item1.id.to_s)
-        expect(input[:name]).to eq("invoice[item_ids][]")
+      it "includes and then removes item from select value", js: true do
+        select_selector = "select[name='invoice[item_ids][]']"
+        expect(find(select_selector, visible: false).value).to include(@item1.id.to_s)
         find(".ss-value-delete").click
         sleep 0.5
-        expect { find(item_id, visible: false) }.to raise_error(Capybara::ElementNotFound)
+        expect(find(select_selector, visible: false).value).not_to include(@item1.id.to_s)
       end
 
       it "does not allow new items", js: true do
         expect_slimselect_options_count_to_eq(3)
-        expect { add_slimselect_option("Not preloaded item") }.to raise_error(Capybara::ElementNotFound)
+        expect do
+          add_slimselect_option("Not preloaded item")
+        end.to raise_error(Capybara::ElementNotFound)
         slimselect_element.send_keys(:escape)
         expect_slimselect_options_count_to_eq(3)
       end
     end
   end
 
-  context "working with active record relations but alias" do
+  context "when working with active record relations but alias" do
     before do
       register_form(Invoice) do |f|
         f.input :other_item_ids, as: :tags, collection: Item.all
@@ -87,14 +87,12 @@ describe "Tags Input", type: :feature do
     context "with added item" do
       before { pick_slimselect_entered_option(@item1.name) }
 
-      it "adds/removes hidden item", js: true do
-        item_id = "#invoice_other_item_ids_#{@item1.id}"
-        input = find(item_id, visible: false)
-        expect(input.value).to eq(@item1.id.to_s)
-        expect(input[:name]).to eq("invoice[other_item_ids][]")
+      it "includes and then removes item from select value", js: true do
+        select_selector = "select[name='invoice[other_item_ids][]']"
+        expect(find(select_selector, visible: false).value).to include(@item1.id.to_s)
         find(".ss-value-delete").click
         sleep 0.5
-        expect { find(item_id, visible: false) }.to raise_error(Capybara::ElementNotFound)
+        expect(find(select_selector, visible: false).value).not_to include(@item1.id.to_s)
       end
     end
   end